src/app/autogen-learning/requirements.txt
# ===== AutoGen 核心包 =====
autogen-core>=0.4.0
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0

# ===== LLM 提供商 SDK =====
openai>=1.0.0

# ===== 环境配置 =====
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ===== 异步支持 =====
aiohttp>=3.9.0

# ===== 开发工具 =====
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.3.0
mypy>=1.0.0
